backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases[sqlite]
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings
python-dotenv
google-generativeai
httpx>=0.25.0
aiosqlite
# Voice processing dependencies (Google-only, lightweight)
deep-translator
gtts
scipy
numpy
SpeechRecognition